Add variables kubeconfig_dir and kubeconfig_file
kubeconfig_dir allows to specify the directory in which to save the
kubeconfig file of the created cluster.

kubeconfig_file allows to specify the exact filename as which to save
the kubeconfig file of the created cluster.

Note: the kubeconfig file is not managed by Terraform. That is, it can
be renamed or moved without this having an effect on the Terraform
state. That also means, when `terraform destroy` is run, the kubeconfig
file is not removed.